add item to collection is now a single url/view, same for remove item from collection. no longer possible to send array of “items” to update-collection-url.
previously, which items were in a collection was modified by sending an update on the collection-url and sending an array of pk in “items” field.

downside with sending the item-array as it should be: can make strange “race” condition if user does:
1. add item pk 3 to a collection which has item [1,2] so far, sends [1,2,3] to server
2. user removes item 1 from collection, sends [2] to server

If request-2 is processed before 1, we get: [1,2,3] as the final result. otherwise we get [2].

This way instead, we just send the specific action to the server so regardless of which request is processed first, we will get the expected [1,3] on both server and front-end.